Revert conversion to long long for KTX branch

+++ source/blender/imbuf/intern/IMB_filetype.h
@@ -40,14 +40,14 @@ typedef struct ImFileType {
 
 	int (*is_a)(unsigned char *buf);
 	int (*is_a_filepath)(const char *name);
-	long long int (*ftype)(const struct ImFileType *type, struct ImBuf *ibuf);
+	int (*ftype)(const struct ImFileType *type, struct ImBuf *ibuf);
 	struct ImBuf *(*load)(unsigned char *mem, size_t size, int flags, char colorspace[IM_MAX_SPACE]);
 	struct ImBuf *(*load_filepath)(const char *name, int flags, char colorspace[IM_MAX_SPACE]);
 	int (*save)(struct ImBuf *ibuf, const char *name, int flags);
 	void (*load_tile)(struct ImBuf *ibuf, unsigned char *mem, size_t size, int tx, int ty, unsigned int *rect);
 
 	int flag;
-	long long int filetype;
+	int filetype;
 	int default_save_role;
 } ImFileType;
